requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cycler==0.10.0
Flask==2.0.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fpdf==1.7.2
greenlet==1.1.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mysql-connector-python==8.0.25
mysqlclient==2.1.0
numpy==1.21.2
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.31
tkhtmlview==0.1.0.post1
urllib3==1.26.6
Werkzeug==2.0.3
WTForms==3.0.1
